Implementación de aapartado de administrador y corrección de errores.

+++ app/src/main/java/com/example/gamehub/models/Game.kt
@@ -1,8 +1,10 @@
 package com.example.gamehub.models
 
+import java.math.BigDecimal
+
 data class Game(
     val id: Int,
     val title: String,
-    val price: Double,
+    val price: BigDecimal,
     val imageUrl: String
 )
